Make viber messsage Id long

+++ DecisionTelecom/Models/ViberMessageReceipt.cs
@@ -8,7 +8,7 @@ namespace DecisionTelecom.Models
         /// Id of the Viber message which status should be got (sent in the last 5 days)
         /// </summary>
         [JsonPropertyName("message_id")]
-        public int ViberMessageId { get; set; }
+        public long ViberMessageId { get; set; }
 
         /// <summary>
         /// Viber message status
